babel-node

+++ ReactTkComponent.js
@@ -1,5 +1,6 @@
 'use strict';
 
+var ReactMultiChild = require('react/lib/ReactMultiChild');
 var ReactTkWindowServer = require('./ReactTkWindowServer');
 
 var assign = require('react/lib/Object.assign');
@@ -10,7 +11,7 @@ function ReactTkComponent() {
   this._rootNodeID = null;
 }
 
-assign(ReactTkComponent.prototype, {
+assign(ReactTkComponent.prototype, ReactMultiChild.Mixin, {
   construct: function(initialElement) {
     this._currentElement = initialElement;
   },
@@ -20,36 +21,53 @@ assign(ReactTkComponent.prototype, {
 
     var widgetType = this._currentElement.type;
     // widgetType needs to be capitalized for Tkinter
-    widgetType = widgetType[0].toUpperCase() + widgetType.slice(1);
+    if (widgetType !== 'container') {
+      widgetType = widgetType[0].toUpperCase() + widgetType.slice(1);
 
-    ReactTkWindowServer.sendCommand({
-      type: 'create',
-      widgetType: widgetType,
-      key: id,
-    });
+      ReactTkWindowServer.sendCommand({
+        type: 'create',
+        widgetType: widgetType,
+        key: id,
+      });
+    }
+
+    if (this._currentElement.props.children) {
+      this.mountChildren(this._currentElement.props.children, transaction, context);
+    }
+
     this.receiveComponent(this._currentElement, transaction, context);
   },
 
   receiveComponent: function(nextElement, transaction, context) {
-    var props = assign({}, nextElement.props);
-    ReactTkWindowServer.sendCommand({
-      type: 'grid',
-      key: this._rootNodeID,
-      properties: props.grid,
-    });
-    delete props.grid;
-    ReactTkWindowServer.sendCommand({
-      type: 'configure',
-      key: this._rootNodeID,
-      properties: props,
-    });
+    if (nextElement.type !== 'container') {
+      var props = assign({}, nextElement.props);
+      ReactTkWindowServer.sendCommand({
+        type: 'grid',
+        key: this._rootNodeID,
+        properties: props.grid,
+      });
+      delete props.grid;
+      ReactTkWindowServer.sendCommand({
+        type: 'configure',
+        key: this._rootNodeID,
+        properties: props,
+      });
+    }
+
+    if (nextElement.props.children) {
+      this.updateChildren(nextElement.props.children, transaction, context);
+    }
   },
 
   unmountComponent: function() {
-    ReactTkWindowServer.sendCommand({
-      type: 'delete',
-      key: this._rootNodeID,
-    });
+    this.unmountChildren();
+
+    if (this._currentElement.type !== 'container') {
+      ReactTkWindowServer.sendCommand({
+        type: 'delete',
+        key: this._rootNodeID,
+      });
+    }
   },
 
   getPublicInstance: function() {
